Object Detection using Ultralytics YOLO v11
Welcome to the HomeObjects-3K object detection using Ultralytics YOLO11 notebook! YOLO11 is the latest version of the YOLO (You Only Look Once) AI models developed by Ultralytics. We hope that the resources in this notebook will help you get the most out of YOLO11.

This notebook serves as a starting point for training the YOLO11 model on home objects detection dataset.